blocked user

+++ src/controllers/properties.controller.ts
@@ -5,10 +5,12 @@ import User from "../models/user.model";
 export const createProperty = async (req: Request, res: Response) => {
   try {
     const user = (req as any).user;
-  
+
     if (!user) return res.status(401).json({ message: "No user found" });
     if (!user.isSeller) {
-      return res.status(403).json({ message: "Only sellers can add properties" });
+      return res
+        .status(403)
+        .json({ message: "Only sellers can add properties" });
     }
     const newProperty = await Product.create({
       ...req.body,
@@ -17,8 +19,8 @@ export const createProperty = async (req: Request, res: Response) => {
     await newProperty.save();
     res.status(201).json({
       message: "Property created successfully, pending approval",
-      property: newProperty
-    })
+      property: newProperty,
+    });
   } catch (error) {
     res.status(500).json({ message: "Error creating property", error });
   }
@@ -31,15 +33,25 @@ export const deleteProperty = async (req: Request, res: Response) => {
       return res.status(404).json({ message: "Property id is required" });
     }
     const sellerId = (req.user as any)._id;
-    const deletedProperty = await Product.findOneAndDelete({
-      _id: id,
-      seller: sellerId, // ensures seller can only delete their own property
-    });
-  
-    if (!deletedProperty) {
-      return res.status(404).json({ message: "Property not found or not owned by this seller" });
+    if (!sellerId) return res.status(404).json({ message: "user not found" });
+    const seller = await User.findById(sellerId);
+    if (seller?.isSeller) {
+      const deletedProperty = await Product.findOneAndDelete({
+        _id: id,
+        seller: sellerId, // ensures seller can only delete their own property
+      });
+
+      if (!deletedProperty) {
+        return res
+          .status(404)
+          .json({ message: "Property not found or not owned by this seller" });
+      }
+      res
+        .status(200)
+        .json({ message: "Property deleted successfully", deletedProperty });
+    } else {
+      res.status(401).json({ message: "Unathourized" });
     }
-    res.status(200).json({ message: "Property deleted successfully", deletedProperty });
   } catch (error) {
     console.error(error);
     return res.status(500).json({ message: "Server error", error });
@@ -60,20 +72,22 @@ export const updateProperty = async (req: Request, res: Response) => {
     if (!seller) {
       return res.status(404).json({ message: "Seller not Found" });
     }
-    if (!seller.isSeller) return res.status(403).json({ message: "Unauthorized" });
+    if (!seller.isSeller)
+      return res.status(403).json({ message: "Unauthorized" });
 
     const updatedProperty = await Product.findOneAndUpdate(
       { _id: id, seller: sellerId },
       { $set: updateFields },
       { new: true }
     );
-      
 
     if (!updatedProperty) {
       return res.status(404).json({ message: "Product not Found" });
     }
 
-    res.status(200).json({ message: "Product updated", product: updatedProperty });
+    res
+      .status(200)
+      .json({ message: "Product updated", product: updatedProperty });
   } catch (error) {
     console.error("Update Property Error:", error);
     res.status(500).json({ message: "Server Error" });
@@ -102,5 +116,18 @@ export const getAllProperties = async (req: Request, res: Response) => {
   } catch (error) {
     console.error("Get All Products Error:", error);
     return res.status(500).json({ message: "Server error" });
+  }
+};
+
+export const getOneProperty = async (req: Request, res: Response) => {
+  try {
+    const sellerId = (req.user as any)._id;
+    const productId = req.params["id"];
+    if (!sellerId) return res.status(404).json({ message: "user not found" });
+    const product = await Product.findOne({ _id: productId , seller: sellerId });
+    if (!product) return res.status(404).json({ message: "product not found" });
+    res.status(200).json(product);
+  } catch (error) {
+    res.status(500).json(error);
   }
 };
